Split the code to classes
